requirements.txt
mcp
llama-index
transformers
sentence-transformers
python-dotenv
llama-index-embeddings-huggingface
einops
quart
requests
qdrant-client
tqdm
httpx